weather_agent/pyproject.toml
[project]
name = "weather-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "donatas.mickevicius",email = "donatas.mickevicius@datatonic.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
]

[tool.poetry.dependencies]
python = "^3.11"
google-cloud-aiplatform = { extras = [
  "adk",
  "agent-engines",
], version = "^1.93.0" }
google-genai = "^1.9.0"
google-adk = "^1.0.0"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
absl-py = "^2.3.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
